Removed Action functionlity from Driver Utils, Other Small Changes
- Removed all Action functionality from Driver Utils
- Moved Driver Utils to a driver_utils folder

- Fixed bug where Pinellas Scraper cannot compare text due to an excessive character.
- Homeowner Search Result now returns the number of removed items when it cleans its list
