+++ guerillo/__main__.py
@@ -1,5 +1,3 @@
-from guerillo.__tests__ import pinellas_tests
-
 # AutoUpdater.run()
 from guerillo.gui import GUI
